% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hello.R
\name{import}
\alias{import}
\title{Import a script or evaluate expression without the Global Environment}
\usage{
import(file, chdir = TRUE)
}
\arguments{
\item{file}{path to source file, or code to evalue. May also be a URL.}

\item{passed}{directly to source function.}
}
\value{
the environment in which the code was evaluated
}
\description{
This function isolates R scripts and code, so they can attach libraries
and load namespaces without affecting other scripts or the Global Environment.
Importantly, this isolation makes the environment returned like an R package,
in that changes to the Global Environment will not affect it.
}
\examples{
module <- import({
 library(MASS)
 gen_data <- function() mvrnorm(10, 0, 1)
})
module$gen_data()                       # can use MASS::mvrnorm
stopifnot(exists("mvrnorm") == FALSE)   # MASS package not attached
mvrnorm <- NULL
module$gen_data()                       # still works
}

