% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_csv_folder.R
\name{import_csv_folder}
\alias{import_csv_folder}
\title{importCsvFolder}
\usage{
import_csv_folder(data_path, rows_to_import = "all", cols_to_import = "all")
}
\arguments{
\item{data_path}{Location of the folder containing the data to import.}

\item{rows_to_import}{First row to import (default = "all").}

\item{cols_to_import}{First column to import (default = "all").}
}
\value{
Return a list of dataframes containing the data from specified
  folder. The dataframes are subset by first row and column specified.
}
\description{
Import all csv files of specific folders to a list of data
  frames.
}
\examples{
\dontrun{importCsvFolder("./datas/myFolder/")}

}
