% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importXlsxFolder.R
\name{importXlsxSep}
\alias{importXlsxSep}
\title{importXlsxFolder}
\usage{
importXlsxSep(dataPath, rowsToImport = "all", colsToImport = "all", ...)
}
\arguments{
\item{dataPath}{Location of the folder containing the data to import.}

\item{rowsToImport}{First row to import (default = "all").}

\item{colsToImport}{First column to import (default = "all").}

\item{...}{Add parameters to "read_excel".}
}
\value{
Return a list of dataframes containing the data from specified
  folder. The dataframes are subset by first row and column specified.
}
\description{
Import all excel files of specific folders to a list of data
  frames.
}
\examples{
\dontrun{importXlsxSep("./datas/myFolder/")}

}
