% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{mergeEnv}
\alias{mergeEnv}
\title{Merge two environments}
\usage{
mergeEnv(e1, e2)
}
\arguments{
\item{e1}{envirnment receiving objects}

\item{e2}{environment giving objects
Adapted from (http://stackoverflow.com/a/26694214/1144523)}
}
\description{
Merge two environments
}

