% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{new_env_chain}
\alias{new_env_chain}
\title{Creates the parent environment used by import}
\usage{
new_env_chain(from)
}
\arguments{
\item{from}{a package on the search path that will be initial parent environment}
}
\description{
This function makes importing possible by creating a chain of environments:
(1) the depends environment, which contains all objects attached via calls to
library or require.
(2) the import environment, this environment has functions to override calls to
the base library and require.
(3) the local environment, which will be returned by import.
}
\examples{
env <- new_env_chain('package:importr')
}

