% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct2datatable.R
\name{struct2dataframe}
\alias{struct2dataframe}
\title{Laod MATLAB structs as R data frames}
\usage{
struct2dataframe(fn, var = c())
}
\arguments{
\item{fn}{A full (character) filename path pointing to the desired .mat file
created by save(). The file must be saved in v6 or before (since R.matlab
can't deal with HDF5-like files yet).}

\item{var}{A character variable name, if a specific variable is desired from 
the .mat file. If no variable is specified, \code{struct2datable} will attempt
to process the first variable listed.}
}
\value{
A data frame will be returned with the struct fields as variables and
  the struct elements as observations
}
\description{
A simple wrapper for readMat for structs that additionally
  converts the resulting list into a data frame.
}
\details{
\code{struct2dataframe} is basically just a replacement for a few
  lines of code, but saves some time and ensures that the struct you clearly
  meant as a data frame for processing or plotting in R actually shows up as 
  a data frame.
}
\examples{
\dontrun{struct2dataframe("~/code/some_struct_v6.mat")}
}
