% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importR.R
\name{import}
\alias{import}
\title{Import a Dataset}
\usage{
import(file)
}
\arguments{
\item{file}{datafile to import}

\item{...parameters}{passed to the import function}
}
\value{
a data frame
}
\description{
This \code{import} function can import data from
delimited text files, simple JSON files, Excel spreadsheets and statistical packages
such as SAS, SPSS, and STATA
}
\details{
The import function is a wrapper for the
\href{https://haven.tidyverse.org/}{haven},
\href{https://readxl.tidyverse.org/}{readxl},
\href{https://cran.r-project.org/web/packages/rjson/index.html}{rjson},
and  \href{https://github.com/r-lib/vroom}{vroom} packages.
}
\note{
Complex nested JSON files will not be imported properly
}
\examples{
\dontrun{
# import a comma delimited file
mydatafram <- import("mydata.csv")

# import a SAS binary file
mydataframe <- import("mydata.sas7ndat")

# import the second worksheet of an excel workbook
mydataframe <- import("mydata.xlsx",sheet=2)

# prompt for a file to import
my dataframe <- import()

}

}
\seealso{
\link[haven]{read_sas},
\link[haven]{read_dta},
\link[haven]{read_spss},
\link[readxl]{read_excel},
\link[rjson]{fromJSON},
\link[vroom]{vroom}
}
