% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{import}
\alias{import}
\title{Import a dataset}
\usage{
import(file, ...)
}
\arguments{
\item{file}{datafile to import}

\item{...}{parameters passed to import function}
}
\value{
a data frame
}
\description{
This \code{import} function can import data from
delimited text files, EXCEL spreadsheets, and
statistical packages such as SAS, SPSS, and Stata
}
\details{
The import is a wrapper for the
\href{https://haven.tidyverse.org/}{haven},
\href{https://readxl.tidyverse.org/}{readxl},
and \href{https://github.com/r-lib/vroom/}{vroom} packages.
}
\examples{
\dontrun{
#import a comma delimited file
myDataFrame <- import("myData.csv")

# import a SAS binary file
myDataFrame <- import("myData.sas7bdat")

#import the second worksheet of an Excel workbook
myDataFrame <- import("myData.xlsx", sheet=2)

#prompt for a file to import
myDataFrame <- import()
}
}
\seealso{
\link[haven]{read_sas},
\link[haven]{read_dta},
\link[haven]{read_spss},
\link[readxl]{read_excel},
\link[vroom]{vroom},
}
