% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import a dataset}
\usage{
import(file, ...)
}
\arguments{
\item{file}{datafile to import. If missing, the user is prompted to select a file interactively.}

\item{...}{parameters passed to the import function. See details.}
}
\value{
a data frame
}
\description{
The import function can import data from delimited text files, excel
spreadsheets, and statistical packages such as SAS, SPSS, and STATA.
}
\details{
put details in here!
}
\examples{
\dontrun{
# import a comma delimited file
my dataframe <- import("mydata.csv")

# import a SAS binary data file
mydataframe <- import("mydata.sas7bdat")

 # import the second worksheet of an Excel workbook
 mydataframe <- import

 # prompt for a file to import
 mydataframe <- import()
 }
}
