% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import a dataset}
\usage{
import(file, ...)
}
\arguments{
\item{file}{datafile to import}

\item{...}{parameters passed to the import function.}
}
\value{
a dataframe
}
\description{
This \code{import} function can import data from
delimited text files, EXCEL spreadsheets, JSON files, and
statistical packages such as SAS, SPSS, and Stata.
}
\details{
The import function is a wrapper for the
\href{https://haven.tidyverse.org/}{haven},
\href{https://readxl.tidyverse.org/}{readxl},
\href{https://cran.r-project.org/web/packages/rjson/index.html}{rjson},
and \href{https://github.com/r-lib/vroom}{vroom} packages.
}
\seealso{
\link[haven]{read_sas},
\link[haven]{read_dta},
\link[haven]{read_spss},
\link[haven]{read_excel},
\link[rjson]{fromJSON},
\link[vroom]{vroom}
}
