% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import a dataset}
\usage{
import(file, ...)
}
\arguments{
\item{file}{datafile to import. If missing, the user
is prompted to select a file interactively.}

\item{...}{parameters passed to the import function. See details below.}
}
\value{
a data frame
}
\description{
The import function can import data from delimited text files,
EXCEL spreadsheets, and statistical packages such as SAS, SPSS, and Stata.
}
\details{
put detail in here!
}
\examples{
\dontrun{
# import comma delimited file
mydataframe <- import("mydata.csv")

# import SAS binary data file
mydataframe <- import("mydata.sas7bdat")

# import the second sheet of an EXCEL workbook
mydataframe <- import("mydata.xlsx", sheet = 2)

# prompt for a file to import
mydataframe <- import()
}
}
