% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import a dataset}
\usage{
import(file, ...)
}
\arguments{
\item{file}{datafile to import. If missing, the user
is prompted to select a file interactively.}

\item{...}{parameters pass to the import functio. See details
below.}
}
\value{
a data frame
}
\description{
the import function can import data from
delimited text files, EXCEL spreadsheets, and
statistical packages such as SAS, SPSS, and Stata.
}
\details{
put details in here!
}
\examples{
\dontrun{
# import a comma delimited file
mydataframe <- import("mydata.csv")

# import a SAS binary datafile
mydataframe <- import("mydata.sas")

# import the second worksheet of an excel workbook
mydataframe <- import("mydata.xlsx", sheet = 2)

# prompt for a file to import
mydataframe <- import()
}
}
